backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
python-decouple==3.8
django-filter==24.3
psycopg[binary]==3.3.2
PyJWT==2.9.0
bleach==6.3.0
Pillow==10.4.0
django-cors-headers==4.7.0
